Fix BPH embed page reader 500 — move PageEditorClient to shared location
The embed page reader was 500ing because importing PageEditorClient
from the [tenant] route tree caused Next.js RSC chunk resolution
failures across route segment boundaries.

Fix: move PageEditorClient to src/components/book/ (shared, no route
tree dependency). The embed page route also uses self-contained data
fetching with findBookByIdOrSlug directly, bypassing tenant resolution.
