+++ src/components/book/PageEditorClient.tsx
